Update non-empty state of picker
