+++ fashion-configuration-defaults/src/commonMain/kotlin/com/aiuta/fashionsdk/configuration/defaults/features/selector/DefaultAiutaImageSelectorFeature.kt
@@ -11,7 +11,6 @@ import com.aiuta.fashionsdk.configuration.features.picker.gallery.photoGallery
 import com.aiuta.fashionsdk.configuration.features.picker.gallery.strings.AiutaImagePickerPhotoGalleryFeatureStrings
 import com.aiuta.fashionsdk.configuration.features.picker.history.dataprovider.AiutaImagePickerUploadsHistoryFeatureDataProviderBuiltIn
 import com.aiuta.fashionsdk.configuration.features.picker.history.strings.AiutaImagePickerUploadsHistoryFeatureStrings
-import com.aiuta.fashionsdk.configuration.features.picker.history.styles.AiutaImagePickerUploadsHistoryFeatureStyles
 import com.aiuta.fashionsdk.configuration.features.picker.history.uploadsHistory
 import com.aiuta.fashionsdk.configuration.features.picker.imagePicker
 import com.aiuta.fashionsdk.configuration.features.picker.model.predefinedModels
@@ -42,7 +41,6 @@ public fun AiutaFeatures.Builder.defaultImagePicker(): AiutaFeatures.Builder = i
         strings = AiutaImagePickerUploadsHistoryFeatureStrings.Default(
             isPredefinedModelAvailable = true,
         )
-        styles = AiutaImagePickerUploadsHistoryFeatureStyles.Default()
         dataProvider = AiutaImagePickerUploadsHistoryFeatureDataProviderBuiltIn
     }
     images = DefaultAiutaImagePickerFeatureImages()
