+++ fashion-configuration/src/commonMain/kotlin/com/aiuta/fashionsdk/configuration/features/tryon/loading/styles/AiutaTryOnLoadingPageFeatureStyles.kt
@@ -1,7 +1,6 @@
 package com.aiuta.fashionsdk.configuration.features.tryon.loading.styles
 
 import androidx.compose.ui.graphics.Color
-import com.aiuta.fashionsdk.configuration.features.styles.AiutaComponentStyle
 
 /**
  * Interface for loading page visual styles.
@@ -10,23 +9,21 @@ import com.aiuta.fashionsdk.configuration.features.styles.AiutaComponentStyle
  * allowing for customization of the appearance of loading UI elements.
  *
  * @property loadingStatusBackgroundGradient Optional gradient colors for the loading status background
- * @property loadingStatusStyle Style for the loading status text and container
  */
 public interface AiutaTryOnLoadingPageFeatureStyles {
     public val loadingStatusBackgroundGradient: List<Color>?
-    public val loadingStatusStyle: AiutaComponentStyle
 
     /**
      * Default implementation of [AiutaTryOnLoadingPageFeatureStyles].
      *
      * Provides standard visual styles for the loading interface,
-     * using a blue gradient background and blurred button style.
+     * using a gradient background.
      */
     public class Default : AiutaTryOnLoadingPageFeatureStyles {
         override val loadingStatusBackgroundGradient: List<Color> = listOf(
-            Color(0xFF4000FF),
-            Color.Transparent,
+            Color.White.copy(alpha = 0.5f),
+            Color(0x00B1B1B1),
+            Color.White.copy(alpha = 0.5f),
         )
-        override val loadingStatusStyle: AiutaComponentStyle = AiutaComponentStyle.BLURRED
     }
 }
