+++ fashion-tryon-compose/src/commonMain/kotlin/com/aiuta/fashionsdk/tryon/compose/ui/internal/screens/selector/content/empty/ImageSelectorScreenEmptyContent.kt
@@ -17,6 +17,8 @@ internal fun ImageSelectorScreenEmptyContent(modifier: Modifier = Modifier) {
         modifier = modifier,
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
+        Spacer(Modifier.height(40.dp))
+
         ImageSelectorScreenEmptyBodyBlock(
             modifier = Modifier.fillMaxWidth().weight(1f),
         )
